Reads all logbooks in a directory. Logbook filenames are assumed to be in the format generated by theCrag, i.e. "climber-logbook-YYYY-MM-DD.csv".

Reads all logbooks in a directory. Logbook filenames are assumed to be in the format generated by theCrag, i.e. "climber-logbook-YYYY-MM-DD.csv".

ReadLogbooks(dir)

Arguments

dir
the directory containing logbook files to read.

Value

a "raw ascents" data frame. This contains one row per ascent, and has the columns:

ascentId: character ID of the ascent

route: character ID of the route

climber: character normalized climber (derived from the filename)

tick: character normalized tick type

grade: integer grade (e.g. Ewbank's grade)

style: integer always 0 (column present for consistency with JSON)

timestamp: integer seconds since UNIX epoch at date of ascent

pitches: integer or NA number of pitches logged